Added cashing to fetch images
Thoughts.
By fetching and storing the imageId when a folder is clicked, we know what image to be displayed at a given time.

When the images are fetched from GoogleDrive they are stored in the ImageContainer, and when a new render of the view are needed, its checking if the image already have been fetched, if the image have been fetched it will look in the cache and display them.

This is done to spare some calls to the GoogleDriveApi